feat(snapshot): add file-first content-addressed disk snapshots
Capture a stopped sandbox's writable upper layer into a self-describing,
content-addressed artifact that can be moved between machines and used
to boot fresh sandboxes. The artifact directory is the source of truth;
the local DB index is a rebuildable cache.

SDK surface (Rust):
- Snapshot, SnapshotHandle, SnapshotBuilder, SnapshotConfig, ExportOpts
- Snapshot::{create, open, get, list, list_dir, remove, reindex,
  export, import}
- SandboxBuilder::from_snapshot() — boot a fresh sandbox from a
  snapshot, mutually exclusive with .image()
- SandboxHandle::snapshot() — snapshot a stopped sandbox

CLI surface:
- msb snapshot create <DEST> --from <SANDBOX> [--label k=v]... [-f]
- msb snapshot {ls, inspect, rm, reindex, export, import}
- msb run --snapshot <PATH_OR_NAME>
- msb snap as a visible alias for msb snapshot

Manifest format (canonical JSON, schema=1):
- Field-order-stable serialization with BTreeMap-sorted labels
- Identity = sha256 of canonical bytes
- format/parent fields present from day one for qcow2 forward-compat
- upper.sha256 verified on every read

Storage:
- Artifacts live under ~/.microsandbox/snapshots/<name>/
  containing manifest.json + upper.ext4
- New snapshot_index table replaces the legacy stub; digest-keyed,
  with parent_digest + child_count for future qcow2 chain ops
- Auto-reindex on first read for artifacts under the configured dir

Sparse-aware copy (microsandbox-utils::copy::fast_copy):
- Tier 1: reflink-copy (clonefile on macOS APFS, FICLONE on Linux
  btrfs/XFS/bcachefs) — zero-copy COW
- Tier 2: SEEK_DATA/SEEK_HOLE walk + copy_file_range for Linux ext4
  fallback; never falls back to dense byte-for-byte copy
- Preserves sparseness on every supported filesystem

Archive format:
- .tar.zst default (sparse-friendly), --plain-tar opt-in
- --with-parents walks the parent chain (no-op in v1)
- --with-image bundles cache/{layers,fsmeta,vmdk} for offline transport
- Magic-byte detection on import

Forward-compat seams for qcow2 backing chains:
- VmConfig.rootfs_upper_spec: UpperSpec { primary, format, backing,
  read_only } — runtime walks the backing chain at attach time
- Manifest.format/parent always emitted; readers reject schema != 1

Tests: 22 new (10 manifest unit, 4 fast_copy with sparseness regression,
8 artifact integration round-trip via tar.zst and plain tar). Existing
DB migration tests updated for the new schema.

New deps: reflink-copy (workspace), astral-tokio-tar +
async-compression (microsandbox crate), hex (microsandbox crate).

Pre-existing clippy warnings in unrelated test code (image registry,
filesystem name validation, network udp_relay/icmp_relay) cleaned up
incidentally to satisfy `-D warnings` after a clippy version bump.

+++ crates/network/lib/proxy.rs
@@ -256,11 +256,11 @@ mod tests {
         // Compression methods length + null
         body.extend_from_slice(&[0x01, 0x00]);
         // Extensions length
-        body.extend_from_slice(&(extensions_total as u16).to_be_bytes());
+        body.extend_from_slice(&extensions_total.to_be_bytes());
         // SNI extension: type 0x0000
         body.extend_from_slice(&[0x00, 0x00]);
-        body.extend_from_slice(&(extension_data_len as u16).to_be_bytes());
-        body.extend_from_slice(&(server_name_list_len as u16).to_be_bytes());
+        body.extend_from_slice(&extension_data_len.to_be_bytes());
+        body.extend_from_slice(&server_name_list_len.to_be_bytes());
         body.push(0x00); // host_name type
         body.extend_from_slice(&host_len.to_be_bytes());
         body.extend_from_slice(host_bytes);
